% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_utils.R
\name{allowsMutations}
\alias{allowsMutations}
\title{Check if a marker allows mutations}
\usage{
allowsMutations(marker)
}
\arguments{
\item{marker}{A \code{marker} object}
}
\value{
Returns TRUE if the \code{mutmat} attribute of the input is non-NULL and differs from the identity matrix.
}
\description{
Check if a marker allows mutations
}
