% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.ped.R
\name{as.data.frame.ped}
\alias{as.data.frame.ped}
\title{Convert ped to data.frame}
\usage{
\method{as.data.frame}{ped}(x, ..., markers)
}
\arguments{
\item{x}{Object of class \code{ped}.}

\item{...}{Further parameters}

\item{markers}{(Optional) Vector of marker indices. By default, all markers
are included.}
}
\description{
Convert a \code{ped} object to a data.frame with columns id, fid, mid and sex. The
output uses the original ID labels, unlike \code{\link[=as.matrix.ped]{as.matrix.ped()}} which uses
internal numeric IDs. The latter is safer for manipulating the pedigree
structure, since the internal IDs are always distinct. (See \code{\link[=setLabels]{setLabels()}}.)
The main use of the data.frame method is printing.
}
