% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.ped.R
\name{as.matrix.ped}
\alias{as.matrix.ped}
\alias{restore_ped}
\title{Convert ped to matrix}
\usage{
\method{as.matrix}{ped}(x, include.attrs = TRUE, ...)

restore_ped(x, attrs = NULL, check = TRUE)
}
\arguments{
\item{x}{a \code{\link{ped}} object. In \code{restore_ped}: A
numerical matrix.}

\item{include.attrs}{a logical indicating if marker annotations and other
info should be attached as attributes. See value.}

\item{\dots}{not used.}

\item{attrs}{a list containing labels and other \code{ped}
info compatible with \code{x}, in the format produced by \code{as.matrix}.
If NULL, the attributes of \code{x} itself are used.}

\item{check}{a logical, forwarded to \code{\link{ped}}. If FALSE, no
checks for pedigree errors are performed.}
}
\value{
For \code{as.matrix}: A matrix with \code{x$NIND} rows.
If \code{include.attrs = TRUE} the matrix has the following attributes:
\itemize{
\item{\code{labels}}{ a list of marker annotations}
\item{\code{famid}}{ the availability vector}
}

For \code{restore_ped}: A \code{ped} object.
}
\description{
Converts a \code{ped} object to a numeric matrix using internal labels,
with additional info neccessary to recreate the original \code{ped}
attached as attributes.
}
\details{
\code{restore_ped} is the reverse of \code{as.matrix.ped}.
}
\examples{

x = setLabels(nuclearPed(1), letters[1:3])

# To examplify the ped -> matrix -> ped trick, we show how to
# reverse the internal ordering of the pedigree.
m = as.matrix(x, include.attrs=TRUE)
m[] = m[3:1, ]

# Must reverse the labels also:
attrs = attributes(m)
attrs$labels = rev(attrs$labels)

# Restore ped:
y = restore_ped(m, attrs=attrs)

# Of course a simpler way is use reorder():
z = reorder(x, 3:1)
stopifnot(identical(y, z))

}
\seealso{
\code{\link{ped}}
}
\author{
Magnus Dehli Vigeland
}
