% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_genotype_import.R
\name{attachGenotypeToPed}
\alias{attachGenotypeToPed}
\title{Attach genotype data to pedigrees}
\usage{
attachGenotypeToPed(ped, references, frequencies, markers = NULL)
}
\arguments{
\item{ped}{a \code{ped} object}

\item{references}{a \code{\link[=data.frame]{data.frame()}} with 4 columns (Id, Marker, Allele 1,
Allele 2)}

\item{frequencies}{a \code{\link[=data.frame]{data.frame()}} with one column per marker and one row
per allele where a cell (allele, marker) contains a value for the frequency}

\item{markers}{a vector of strings with the marker names that should be
attached to the pedigree. Defaults to \code{NULL} in which case all information
available in \code{references} will be attached to the pedigree}
}
\value{
a \code{ped} object with the attached markers
}
\description{
Reads genotype data from a reference dataframe consisting of four columns:
Id, Marker, Allele 1 and Allele 2. Creates one marker per row and attaches
them to the pedigree in the correct format. Allele denomination and frequency
data are read from the frequency dataframe which must contain a table with
one allele per row and one marker per column, where a cell (allele, marker)
contains the frequency of that allele in the given population. Only positive
numerical values are passed to the pedigree (alleles with frequencies 0 or NA
are ignored). Optionally, the user can choose to only attach some of the
markers in the reference file by passing a list of which ones to attach.
}
\details{
If the pedigree has genotype information already attached to it, that info is
left intact.
}
\seealso{
\code{\link[=as.ped]{as.ped()}}
}
\author{
Elias Hernandis \href{mailto:eliashernandis@gmail.com}{eliashernandis@gmail.com}
}
