% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped.R
\name{checkped}
\alias{checkped}
\title{Pedigree errors}
\usage{
checkped(x)
}
\arguments{
\item{x}{object of class \code{ped}.}
}
\value{
If no errors are detected, the function returns NULL invisibly.
  Otherwise, messages describing the errors are printed to the screen and an
  error is raised.
}
\description{
Check a \code{ped} object for pedigree errors.
}
