% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_stats.R
\name{countHomozygous}
\alias{countHomozygous}
\title{Count homozygous markers}
\usage{
countHomozygous(x, ids = NULL)
}
\arguments{
\item{x}{A pedigree (ped object) with marker data.}

\item{ids}{A vector of individual ID labels. If NULL (default) all
individuals are included.}
}
\value{
A named integer vector with the counts of homozygous markers for
each individual in \code{ids}.
}
\description{
Counts the number of homozygous genotypes for each individual.
}
\examples{
x = nuclearPed() |> addMarker(`2` = "1/1", `3` = "1/2")
countHomozygous(x)

}
