% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{has_common_ancestor}
\alias{has_unbroken_loops}
\alias{has_selfing}
\alias{has_inbred_founders}
\alias{has_parents_before_children}
\alias{parents_before_children}
\alias{is_Xmarker}
\alias{restore_ped}
\alias{validate_ped}
\title{Deprecated functions}
\usage{
has_common_ancestor(x)

has_unbroken_loops(x)

has_selfing(x)

has_inbred_founders(x)

has_parents_before_children(x)

parents_before_children(x)

is_Xmarker(x)

restore_ped(x)

validate_ped(x)
}
\arguments{
\item{x}{object}
}
\description{
These functions have been renamed (from "snake_case" to "camelCase") and will eventually be removed.
}
