% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_distribute.R
\name{distributeMarkers}
\alias{distributeMarkers}
\title{Attach identical markers evenly distributed across the genome}
\usage{
distributeMarkers(
  x,
  n = NULL,
  dist = NULL,
  chromLen = NULL,
  alleles = 1:2,
  afreq = NULL,
  prefix = NULL
)
}
\arguments{
\item{x}{A ped object.}

\item{n}{The total number of markers. Either this or \code{dist} must be NULL.}

\item{dist}{A positive number; the distance (in megabases) between markers.}

\item{chromLen}{A numeric vector indicating chromosome lengths (in Mb). By
default, the lengths of the human chromosomes 1-22 are used, as returned by
\code{sapply(ibdsim2::loadMap("decode"), ibdsim2::physRange)}.}

\item{alleles, afreq}{Passed onto \code{\link[=marker]{marker()}}.}

\item{prefix}{A string used as prefix for marker names. The default (NULL)
gives no names.}
}
\value{
A ped object similar to \code{x}.
}
\description{
Attach identical markers evenly distributed across the genome
}
\examples{
x = distributeMarkers(nuclearPed(), n = 10, prefix = "M")
getMap(x)

}
