% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_subgroups.R
\name{extractSingletons}
\alias{extractSingletons}
\title{Extract singletons from pedigree}
\usage{
extractSingletons(x, ids = NULL, keepFI = TRUE)
}
\arguments{
\item{x}{A \code{ped} object or a list of such.}

\item{ids}{A vector of ID labels (coercible to character).}

\item{keepFI}{A logical indicating if founder inbreeding should be preserved,
if present.}
}
\value{
A list of singletons.
}
\description{
Extract singletons from pedigree
}
\examples{

x = nuclearPed() |> addMarker(geno = c("1/1", NA, "1/2"))

# Extract father and child
extractSingletons(x, ids = c(1,3))

}
