% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_accessors.R
\name{famid}
\alias{famid}
\alias{famid.ped}
\alias{famid<-}
\alias{famid<-.ped}
\title{Family identifier}
\usage{
famid(x, ...)

\method{famid}{ped}(x, ...)

famid(x, ...) <- value

\method{famid}{ped}(x, ...) <- value
}
\arguments{
\item{x}{A \code{ped} object}

\item{...}{(Not used)}

\item{value}{The new family ID, which must be (coercible to) a character
string.}
}
\description{
Functions for getting or setting the family ID of a \code{ped} object.
}
\examples{
x = nuclearPed(1)
famid(x) # empty string

famid(x) = "trio"
famid(x)

}
