% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_founder-inbreeding.R
\name{founder_inbreeding}
\alias{founder_inbreeding}
\alias{founder_inbreeding<-}
\title{Inbreeding coefficients of founders}
\usage{
founder_inbreeding(x, ids, named = FALSE)

founder_inbreeding(x, ids) <- value
}
\arguments{
\item{x}{A \code{ped} object}

\item{ids}{Any subset of \code{founders(x)}. If \code{ids} is missing in
\code{founder_inbreeding()}, it is set to \code{founders(x)}.}

\item{named}{A logical: If TRUE, the output vector is named with the ID
labels}

\item{value}{A numeric of the same length as \code{ids}, entries in the interval
\code{[0, 1]}. If the vector is named, then the names are interpreted as ID
labels of the founders whose inbreeding coefficients should be set. In this
case, the \code{ids} argument should not be used. (See examples.)}
}
\value{
For \code{founder_inbreeding}, a numeric vector of the same length as
\code{ids}, containing the founder inbreeding coefficients.

For \code{founder_inbreeding<-} the updated \code{ped} object is returned.
}
\description{
Functions to get or set inbreeding coefficients for the pedigree founders.
}
\examples{
x = nuclearPed(1)
founder_inbreeding(x, ids = '1') = 1
founder_inbreeding(x, named = TRUE)

# Setting all founders at once (replacement value is recycled)
founder_inbreeding(x, ids = founders(x)) = 0.5
founder_inbreeding(x, named = TRUE)

# Alternative syntax, using a named vector
founder_inbreeding(x) = c('1'=0.1, '2'=0.2)
founder_inbreeding(x, named = TRUE)

}
