% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_allelematrix.R
\name{getAlleles}
\alias{getAlleles}
\alias{setAlleles}
\title{Allele matrix manipulation}
\usage{
getAlleles(x, ids = NULL, markers = NULL)

setAlleles(x, ids = NULL, markers = NULL, alleles)
}
\arguments{
\item{x}{A \code{ped} object or a list of such}

\item{ids}{A vector of ID labels. If NULL (default) all individuals are
included.}

\item{markers}{A vector of indices or names of markers attaches to \code{x}. If
NULL (default) all markers are included.}

\item{alleles}{A character (or coercible to character) matrix of dimensions
\code{length(x) * nMarkers(x)}}
}
\value{
A character matrix of dimensions \code{length(x) * nMarkers(x)}
}
\description{
Functions for getting and setting the genotypes of multiple
individuals/markers simultaneously
}
\examples{
x = nuclearPed(1)
m1 = marker(x, `2` = 1:2, alleles = 1:2, name="m1")
m2 = marker(x, `3` = 2, alleles = 1:2, name="m2")
x = setMarkers(x, list(m1, m2))

mat1 = getAlleles(x)
mat2 = getAlleles(x, ids=2:3, markers = "m2")
stopifnot(identical(mat1[2:3, 3:4], mat2))

# Remove all genotypes
y = setAlleles(x, alleles = 0)
y

# Setting a single genotype
z = setAlleles(y, ids="1", marker = "m2", alleles = 1:2)

# Alternative: In-place modification with `genotype()`
genotype(y, id = "1", marker = "m2") = 1:2
stopifnot(identical(y,z))

### Manipulation of pedlist objects
s = transferMarkers(x, singleton("s"))
peds = list(x, s)

getAlleles(peds)

setAlleles(peds, ids="s", marker="m1", alleles=1:2)

}
