% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_database.R
\name{getFrequencyDatabase}
\alias{getFrequencyDatabase}
\title{Allele frequency database}
\usage{
getFrequencyDatabase(x, markers = seq_len(nMarkers(x)))
}
\arguments{
\item{x}{A \code{ped} object}

\item{markers}{A character vector (with marker names) or a numeric vector
(with marker indices)}
}
\value{
A data.frame with allele labels as now names, and marker labels as
column names
}
\description{
Extract marker allele frequencies as a data.frame, sorted according to the
'allelic ladder'
}
\examples{
loc1 = list(name = "m1", alleles = 3:4, afreq = c(.1, .9))
loc2 = list(name = "m2", alleles = c("1", "10.2", "3"))
x = setMarkers(singleton(1), locus = list(loc1, loc2))
getFrequencyDatabase(x)

}
