% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMap.R
\name{getMap}
\alias{getMap}
\title{Tabulate marker positions}
\usage{
getMap(
  x,
  markers = seq_len(nMarkers(x)),
  pos = c("cm", "mb"),
  na.action = 0,
  verbose = TRUE
)
}
\arguments{
\item{x}{An object of class \code{ped}.}

\item{markers}{A numeric of indices.}

\item{pos}{Which unit should be used? Either "cm" (centiMorgan) or "mb" (megabytes).}

\item{na.action}{Either 0 (default), 1 or 2.}

\item{verbose}{A logical.}
}
\value{
A \code{data.frame}.
}
\description{
Tabulate marker positions
}
