% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_utils.R
\name{getSex}
\alias{getSex}
\title{Return the sex of specified pedigree members.}
\usage{
getSex(x, labels)
}
\arguments{
\item{x}{A \code{ped} object.}

\item{labels}{A character vector (or coercible to one) containing ID labels.}
}
\value{
An integer vector of the same length as \code{labels}, with entries 0
(unknown), 1 (male) or 2 (female).
}
\description{
Return the sex of specified pedigree members.
}
\examples{
x = nuclearPed(1)
stopifnot(getSex(x, 1) == 1)

}
