% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_utils.R
\name{hasCA}
\alias{hasCA}
\title{Pairwise common ancestors}
\usage{
hasCA(x)
}
\arguments{
\item{x}{a \code{\link[=ped]{ped()}} object.}
}
\description{
Computes a matrix A whose entry \code{A[i,j]} is TRUE if pedigree members i and j have a common ancestor, and FALSE otherwise.
}
\examples{

x = fullSibMating(3)
A = hasCA(x)
stopifnot(A[1,1], !A[1,2], all(A[3:8, 3:8]))

}
