% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_loops.R
\name{inbreedingLoops}
\alias{inbreedingLoops}
\title{Inbreeding loops}
\usage{
inbreedingLoops(x)
}
\arguments{
\item{x}{A \code{ped} object.}
}
\value{
A list of inbreeding loops, where each loop is a list with entries
\code{top}, \code{bottom}, \code{pathA} and \code{pathB}. The \code{top} and \code{bottom} entries give
the IDs of the individuals at the top and bottom of the loop, while \code{pathA}
and \code{pathB} give the IDs of the individuals along the two paths from top to
bottom (excluding top and bottom themselves).
}
\description{
Returns a list of all inbreeding loops in the pedigree. Note that this may be large
(and slow) in heavily inbred pedigrees. For example, \code{fullSibMating(10)} has 8144
loops. For finding loop breakers and breaking loops, see \code{findLoopBreakers()} and
\code{breakLoops()}, which are much more efficient for large pedigrees, and do not require
enumerating all loops.
}
\examples{
x = cousinPed(1, child = TRUE)
loops = inbreedingLoops(x)

}
\seealso{
\code{\link[=findLoopBreakers]{findLoopBreakers()}}, \code{\link[=breakLoops]{breakLoops()}}
}
