% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped.R
\name{internalID}
\alias{internalID}
\title{Internal pedigree order}
\usage{
internalID(x, labels)
}
\arguments{
\item{x}{A \code{ped} object.}

\item{labels}{A character vector (or coercible to one) of original ID labels.}
}
\value{
A numeric vector
}
\description{
Return the internal indices of pedigree members.
}
\examples{
x = nuclearPed(1)
internalID(x, 3)

}
