% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_attach.R
\name{marker_attach}
\alias{marker_attach}
\alias{setMarkers}
\alias{addMarkers}
\title{Attach markers to pedigrees}
\usage{
setMarkers(x, m, annotations = NULL)

addMarkers(x, m, annotations = NULL)
}
\arguments{
\item{x}{A \code{ped} object}

\item{m}{Either a single \code{marker} object, a list of \code{marker} objects, or a
data.frame or matrix.}

\item{annotations}{A list of marker annotations.}
}
\value{
A \code{ped} object.
}
\description{
In many applications it is useful to \emph{attach} markers to their associated
\code{ped} object. In particular for bigger projects with many markers, this makes
it easier to manipulate the dataset as a unit. The function \code{setMarkers()}
replaces all existing markers with the supplied ones, while \code{addMarkers()}
appends them to any existing ones.
}
\examples{
x = singleton(1)
m1 = marker(x, '1'=1:2)
m2 = marker(x, '1'='a')

x = setMarkers(x, m1)
x = addMarkers(x, m2)
x

# Reversing the order of the markers
x = setMarkers(x, list(m2, m1))
x

}
