% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_mask.R
\name{maskPed}
\alias{maskPed}
\alias{unmaskPed}
\title{Mask and unmask pedigree datasets}
\usage{
maskPed(x, seed = NULL)

unmaskPed(x, keys)
}
\arguments{
\item{x}{A \code{ped} object or a list of such.}

\item{seed}{An optional seed for the random number generator.}

\item{keys}{A list with entries \code{ids}, \code{markers}, \code{alleles}.}
}
\value{
An object similar to \code{x} but with replaced ID labels, marker names
and allele labels.
}
\description{
The \code{maskPed()} function replaces the individual IDs, marker names and allele
labels with generic sequences like 1, 2, ... For markers with stepwise
mutation models, the allelic ladder is simply translated to start at 1,
thereby preserving the intra-allelic differences.
}
\details{
It should be noted that when the masking procedure is applied to a dataset
using publicly available frequency databases, the result cannot be considered
to be fully anonymised. (In theory, one could deduce the original marker
names and alleles from the frequencies.)
}
\examples{
x = nuclearPed(father = "fa", mother = "mo", children = "ch") |>
  addMarker(name = "myMarker", ch = "b/c", afreq = c(a=0.2, b=0.3, c=0.5)) |>
  setMutmod(model = "proportional", rate = 0.01)

# Mask
y = maskPed(x, seed = 1729)

# Unmask
z = unmaskPed(y$maskedPed, y$keys)
stopifnot(identical(x, z))

# With stepwise model
x2 = x |>
  addMarker(name = "mySTR", ch = "7.2/8.2",
            alleles = c("7", "7.2", "8", "8.2")) |>
  setMutmod(marker = 2, model = "stepwise", rate = 0.1, rate2 = 1e-6,
            range = 0.1)

y2 = maskPed(x2, seed = 1729)

z2 = unmaskPed(y2$maskedPed, y2$keys)

stopifnot(identical(x2, z2))

# Check likelihoods with pedprobr:
# stopifnot(setequal(likelihood(x2), likelihood(y2$maskedPed)))

}
