% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePed.R
\name{mergePed}
\alias{mergePed}
\title{Merge two pedigrees}
\usage{
mergePed(x, y, ...)
}
\arguments{
\item{x, y}{\code{\link{ped}} objects}

\item{...}{further arguments passed along to \code{\link{ped}}, e.g.
\code{famid}, \code{check} and \code{reorder}.}
}
\value{
A \code{ped} object.
}
\description{
This function merges two ped objects, joining them at the individuals
with equal ID labels. This is especially useful for building 'top-heavy'
pedigrees. Only ped objects without marker data are supported.
}
\examples{

# Creating a trio where each parent have first cousin parents.
# (Alternatively, this could be built using many calls to addParents().)

x = cousinsPed(1)
x = addChildren(x, father=5, mother=8, nch=1, ids=9)
x = addChildren(x, father=9, mother=10, nch=1, ids=11)

y = relabel(cousinsPed(1), 101:108)
y = addChildren(y, father=105, mother=108, nch=1, sex=2, id=10)
y = addChildren(y, father=9, mother=10, nch=1, id=11)

# Joining x and y at the common individuals 9,10,11:
z = mergePed(x,y)

# plot all three pedigrees
par(mfrow=c(1,3))
plot(x); plot(y); plot(z)

}
\author{
Magnus Dehli Vigeland
}
