% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_mutations.R
\name{mutationMatrix}
\alias{mutationMatrix}
\title{Mutation matrices}
\usage{
mutationMatrix(alleles, model = c("equal", "proportional", "random"),
  rate)
}
\arguments{
\item{alleles}{A character vector (or coercible to character) with allele labels}

\item{model}{A string: either "equal", "proportional" or "random"}

\item{rate}{A number between 0 and 1}
}
\value{
A square matrix whose \code{dimnames} are given by the \code{alleles}.
}
\description{
Construct mutation matrices
}
\details{
At the moment only \code{model = equal} is implemented
}
\examples{
mutationMatrix(alleles = 1:3, model = "equal", rate = 0.05)

}
