% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_utils.R
\name{nAlleles}
\alias{nAlleles}
\title{Number of marker alleles}
\usage{
nAlleles(m)
}
\arguments{
\item{m}{An object of class \link{marker}.}
}
\value{
A positive integer.
}
\description{
Number of marker alleles
}
\examples{
x = nuclearPed(1)
m = marker(x)
nAlleles(m)

}
