% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped.R
\name{newPed}
\alias{newPed}
\title{Internal ped constructor}
\usage{
newPed(ID, FIDX, MIDX, SEX, FAMID, detectLoops = TRUE)
}
\arguments{
\item{ID}{A character vector.}

\item{FIDX}{An integer vector.}

\item{MIDX}{An integer vector.}

\item{SEX}{An integer vector.}

\item{FAMID}{A string.}

\item{detectLoops}{A logical.}
}
\value{
A \code{ped} object.
}
\description{
This is the internal constructor of \code{ped} objects. It does not do any
validation of input other than simple type checking. In particular it should
only be used in programming scenarios where it is known that the input is a
valid, connected pedigree. End users are recommended to use the regular
constructor \code{\link[=ped]{ped()}}.
}
\details{
See \code{\link[=ped]{ped()}} for details about the input parameters.
}
\examples{

newPed("a", 0L, 0L, 1L, "")

}
