% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped.R
\name{parents_before_children}
\alias{parents_before_children}
\alias{has_parents_before_children}
\title{Standard pedigree order}
\usage{
parents_before_children(x)

has_parents_before_children(x)
}
\arguments{
\item{x}{a \code{\link[=ped]{ped()}} object}
}
\description{
Reorder a \code{ped} object so parents come before their children.
}
\examples{
x = reorder(nuclearPed(1), 3:1)
x
parents_before_children(x)

}
