% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped.R
\name{ped}
\alias{ped}
\alias{singleton}
\title{Pedigree construction}
\usage{
ped(id, fid, mid, sex, famid = NULL, reorder = TRUE, check = TRUE,
  verbose = FALSE)

singleton(id, sex = 1, famid = NULL)
}
\arguments{
\item{id}{a vector (numeric or character) of individual ID labels.}

\item{fid}{a vector of the same length as \code{id}, containing the labels of the fathers.
In other words \code{fid[i]} is the father of \code{id[i]}, or 0 if \code{id[i]} is a founder.}

\item{mid}{a vector of the same length as \code{id}, containing the labels of the mothers.
In other words \code{mid[i]} is the mother of \code{id[i]}, or 0 if \code{id[i]} is a founder.}

\item{sex}{a numeric of the same length as \code{id}, describing the genders of the individuals
(in the same order as \code{id}.) Each entry must be either 1 (=male), 2 (=female) or 0 (=unknown).}

\item{famid}{a character of length 1}

\item{reorder}{a logical. If TRUE, the pedigree is reordered so that all
parents preceeede their children.}

\item{check}{a logical. If TRUE, \code{\link[=checkped]{checkped()}} is run on the
pedigree before it is returned.}

\item{verbose}{a logical.}

\item{...}{further arguments}
}
\value{
A \code{ped} object, which is essentially a list with the following
entries:
\describe{
\item{ID}{A numerical vector with the internal IDs, which are always 1,2,3,...,N.}
\item{FID}{A numerical vector indicating the internal IDs of the fathers.}
\item{MID}{A numerical vector indicating the internal IDs of the mothers.}
\item{SEX}{A numerical vector with gender codes. Unless the pedigree is reordered,
this equals the input argument \code{sex}.}
\item{FAMID}{The family ID.}
\item{LABELS}{A character vector containing the original id labels.
Unless the pedigree has been reordered, this equals the input argument \code{id}.}
\item{hasLoops}{A logical: TRUE if the pedigree is inbred.}
\item{loop_breakers}{A matrix with loop breaker ID's in the first
column and their duplicates in the second column. All entries refer to the internal IDs.
This is usually set by \code{\link[=breakLoops]{breakLoops()}}.}
}
}
\description{
Basic construction of \code{ped} objects. Utility functions for creating many
common pedigree structures are described in \link{ped_create}.
}
\details{
Internally, this happens: ...

A singleton is a special \code{ped} object whose pedigree contains 1
individual. The class attribute of a singleton is \code{c('singleton', 'ped')}
}
\examples{
x = ped(id=1:3, fid=c(0,0,1), mid=c(0,0,2), sex=c(1,2,1))
y = singleton('NN', sex=2, famid="SINGLETON GIRL")

}
\seealso{
\link{ped_create}, \link{ped_modify}, \link{ped_subsets}
}
\author{
Magnus Dehli Vigeland
}
