% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_complex.R
\name{ped_complex}
\alias{ped_complex}
\alias{doubleCousins}
\alias{doubleFirstCousins}
\alias{quadHalfFirstCousins}
\alias{fullSibMating}
\alias{halfSibStack}
\title{Complex pedigree structures}
\usage{
doubleCousins(degree1, degree2, removal1 = 0, removal2 = 0,
  half1 = FALSE, half2 = FALSE, child = FALSE)

doubleFirstCousins()

quadHalfFirstCousins()

fullSibMating(n)

halfSibStack(n)
}
\arguments{
\item{degree1, degree2, removal1, removal2}{Nonnegative integers.}

\item{half1, half2}{Logicals, indicating if the fathers (resp mothers) should
be full or half cousins.}

\item{child}{A logical: Should a child be added to the double cousins?}

\item{n}{A positive integer indicating the number of crossings.}
}
\value{
A \code{\link{ped}} object.
}
\description{
These functions create certain classes of pedigrees that are not
straightforward to construct starting from the simple structures described in
\link{ped_basic}.
}
\details{
The function \code{doubleCousins} returns a pedigree linking two individuals who
are simultaneous paternal and maternal cousins. More precisely, they are:
\itemize{
\item paternal (full or half) cousins of type (\code{degree1}, \code{removal1})
\item maternal (full or half) cousins of type (\code{degree2}, \code{removal2}).
}

For convenience, a wrapper \code{doubleFirstCousins} is provided for the most
common case, double first cousins.

\code{quadHalfFirstCousins} produces a pedigree with quadruple half first cousins.

\code{fullSibMating} crosses full sibs consecutively \code{n} times.

\code{halfSibStack} produces a breeding scheme where the two individuals in the
final generation are simultaneous half k'th cousins, for each \code{k=0,...,n-1}.
}
\examples{

# Consecutive brother-sister matings.
fullSibMating(2)

# Simultaneous half siblings and half first cousins
halfSibStack(2)

# Double first cousins
x = doubleFirstCousins()
plot(x)

# Quadruple half first cousins
QHFC = quadHalfFirstCousins()
# plot(QHFC) # Weird plotting behaviour for this pedigree.

# A double half cousins pedigree with inbred founders,
# with the same identity coefficients as QHFC above.
# (Requires the `ribd` package.)
\dontrun{
x = doubleCousins(degree1 = 1, removal1 = 1, half1 = TRUE,
                  degree2 = 0, removal2 = 1, half2 = TRUE)
founderInbreeding(x, 1) = 3-2*sqrt(2)
founderInbreeding(x, 4) = .5 * sqrt(2)
j1 = ribd::condensedIdentity(x, leaves(x))
j2 = ribd::condensedIdentity(QHFC, leaves(QHFC))
stopifnot(identical(j1, j2))
}
}
\seealso{
\link{ped_basic}
}
