% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_accessors.R
\name{ped_deprecated}
\alias{ped_deprecated}
\alias{setLabels}
\title{Deprecated functions}
\usage{
setLabels(x, labels)
}
\arguments{
\item{x}{NA}

\item{labels}{NA}
}
\description{
These will soon be removed.
}
