% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_utils.R
\name{ped_utils}
\alias{ped_utils}
\alias{has_unbroken_loops}
\alias{has_inbred_founders}
\alias{labels.ped}
\alias{pedsize}
\title{Pedigree utilities}
\usage{
has_unbroken_loops(x)

has_inbred_founders(x)

\method{labels}{ped}(object, ...)

pedsize(x)
}
\arguments{
\item{x, object}{A \code{ped} object}
}
\description{
Various utility functions for \code{ped} objects
}
\examples{
x = nuclearPed(1)
stopifnot(pedsize(x)==3)

}
