% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedCreate.R
\name{pedCreate}
\alias{pedCreate}
\alias{nuclearPed}
\alias{cousinsPed}
\alias{halfCousinsPed}
\title{Create simple pedigrees}
\usage{
nuclearPed(nch, sex = 1, father = "1", mother = "2",
  children = as.character(seq.int(3, length.out = nch)))

cousinsPed(degree, removal = 0, degree2 = NULL, child = FALSE)

halfCousinsPed(degree, removal = 0, degree2 = NULL, child = FALSE)
}
\arguments{
\item{nch}{A positive integer indicating the number of offspring. If NULL, it
is taken to be the \code{length(children)}}

\item{sex}{A numeric vector of length \code{nch} (recycled if shorter)
encoding the genders of the children (0=unknown, 1=male, 2=female).}

\item{father}{The label of the father.}

\item{mother}{The label of the father.}

\item{children}{A character of length \code{nch}, with labels of the
children.}

\item{degree, degree2}{Non-negative integers, indicating the degree of
cousin-like relationships: 0=siblings, 1=first cousins; 2=second cousins,
a.s.o. See Details and Examples.}

\item{removal}{Non-negative integers, indicating the removals of cousin-like
relationships. See Details and Examples.}

\item{child}{A logical: Should an inbred child be added to the two cousins?}
}
\value{
A \code{\link[=ped]{ped()}} object.
}
\description{
These are utility functions for creating some common pedigree structures as
\code{ped} objects. Use \code{\link[=swapSex]{swapSex()}} to change the gender of
pedigree members.
}
\details{
The call \code{cousinsPed(degree=n, removal=k)} creates a pedigree with two
n'th cousins, k times removed. By default, removals are added on the right
side. To override this, the parameter \code{degree2} can be used to indicate
explicitly the number of generations on the right side of the pedigree. When
\code{degree2} is given \code{removal} is ignored. (Similarly for
\code{halfCousinsPed}.)
}
\examples{

# A nuclear family with 2 boys and 3 girls,
x = nuclearPed(5, sex=c(1,1,2,2,2))

# Half sibs:
halfCousinsPed(degree=0)

# Grand aunt:
cousinsPed(degree=0, removal=2)

# Second cousins once removed.
cousinsPed(degree=2, removal=1)

# Same, but with the 'removal' on the left side.
cousinsPed(degree=3, degree2=2)

# A child of first cousin parents.
cousinsPed(degree=1, child=TRUE)

}
\seealso{
\code{\link[=swapSex]{swapSex()}}, \code{\link[=removeIndividuals]{removeIndividuals()}},
\code{\link[=addChildren]{addChildren()}}, \code{\link[=relabel]{relabel()}},
\code{\link[=doubleCousins]{doubleCousins()}},
}
