% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedModify.R
\name{pedModify}
\alias{pedModify}
\alias{swapSex}
\alias{addChildren}
\alias{addOffspring}
\alias{addSon}
\alias{addDaughter}
\alias{addParents}
\alias{removeIndividuals}
\alias{branch}
\alias{relabel}
\title{Modify the pedigree of 'ped' objects}
\usage{
swapSex(x, ids, verbose = TRUE)

addChildren(x, father = NULL, mother = NULL, nch = 1, sex = 1,
  ids = NULL, verbose = TRUE)

addOffspring(x, father = NULL, mother = NULL, nch = 1, sex = 1,
  ids = NULL, verbose = TRUE)

addSon(x, parent, id = NULL, verbose = TRUE)

addDaughter(x, parent, id = NULL, verbose = TRUE)

addParents(x, id, father = NULL, mother = NULL, verbose = TRUE)

removeIndividuals(x, ids, verbose = TRUE)

branch(x, id)

relabel(x, new, old = x$LABELS)
}
\arguments{
\item{x}{A \code{\link[=ped]{ped()}} object}

\item{verbose}{A logical: Verbose output or not.}

\item{father, mother}{Integers indicating the IDs of parents. If missing, a
new founder individual is created (whose ID will be 1+the largest ID already
in the pedigree).}

\item{nch}{A single integer indicating the number of children to be
created. Default: 1.}

\item{sex}{Integer vectors indicating the genders
of the offspring to be created (recycled if less than \code{nch}
elements).}

\item{parent}{Integer ID of any pedigree member, which will be the father or
mother (depending on its gender) of the new child.}

\item{id, ids}{Individual ID label(s). In \code{addOffspring} the (optional)
\code{ids} argument is used to specify ID labels for the offspring to be
created.}

\item{new}{a numeric containing new labels to replace those in \code{old}.}

\item{old}{a numeric containing ID labels to be replaced by those in
\code{new}. If missing, \code{old} is set to \code{x$LABELS}.}
}
\value{
The modified \code{ped} object.
}
\description{
Functions to modify the pedigree of a 'ped' object.
}
\details{
When removing an individual, all descendants are also removed as well as
founders remaining without offspring.

The \code{branch()} function extracts the pedigree subset consisting of all
descendants of \code{id}, including \code{id} itself and all relevant
spouses.
}
\examples{

x = nuclearPed(1)

# To see the effect of each command below, use plot(x) in between.
x = addSon(x, 3)
x = addParents(x, id=4, father=6, mother=7)
x = swapSex(x, c(1,4))
x = removeIndividuals(x, 4)

}
\seealso{
\code{\link[=ped]{ped()}}, \code{\link[=nuclearPed]{nuclearPed()}}
}
\author{
Magnus Dehli Vigeland
}
