% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedParts.R
\name{pedParts}
\alias{pedParts}
\alias{founders}
\alias{nonfounders}
\alias{children}
\alias{offspring}
\alias{spouses}
\alias{unrelated}
\alias{leaves}
\alias{parents}
\alias{grandparents}
\alias{siblings}
\alias{cousins}
\alias{nephews_nieces}
\alias{ancestors}
\alias{descendants}
\title{Pedigree subsets}
\usage{
founders(x, internal = FALSE)

nonfounders(x, internal = FALSE)

children(x, id, internal = FALSE)

offspring(x, id, internal = FALSE)

spouses(x, id, internal = FALSE)

unrelated(x, id, internal = FALSE)

leaves(x, internal = FALSE)

parents(x, id, internal = FALSE)

grandparents(x, id, degree = 2, internal = FALSE)

siblings(x, id, half = NA, internal = FALSE)

cousins(x, id, degree = 1, removal = 0, half = NA, internal = FALSE)

nephews_nieces(x, id, removal = 1, half = NA, internal = FALSE)

ancestors(x, id, internal = FALSE)

descendants(x, id, internal = FALSE)
}
\arguments{
\item{x}{A \code{\link{ped}} object.}

\item{internal}{A logical indicating whether 'id' refers to the internal order.}

\item{id}{A single ID label (numeric or character).}

\item{degree, removal}{Non-negative integers.}

\item{half}{a logical or NA. If TRUE (resp FALSE), only half (resp. full)
siblings/cousins/nephews/nieces are returned. If NA, both categories are
inclucded.}
}
\value{
For \code{ancestors(x,id)}, a vector containing the ID's of all
ancestors of the individual \code{id}.  For \code{descendants(x,id)}, a
vector containing the ID's of all descendants (i.e. children, grandchildren,
a.s.o.) of individual \code{id}.

The functions \code{cousins}, \code{grandparents}, \code{nephews_nieces},
\code{children}, \code{parents}, \code{siblings}, \code{spouses},
\code{unrelated}, each returns an integer vector containing the ID's of all
pedigree members having the specified relationship with \code{id}.

For \code{leaves}, a vector of the ID labels of pedigree members without
children.
}
\description{
Utility functions for 'ped' objects, mainly for extracting various
pedigree information.
}
\examples{

x = ped(id=2:9, fid=c(0,0,2,0,4,4,0,2), mid=c(0,0,3,0,5,5,0,8),
        sex=c(1,2,1,2,1,2,2,2))
stopifnot(setequal(spouses(x, 2), c(3,8)),
          setequal(children(x, 2), c(4,9)),
          setequal(descendants(x, 2), c(4,6,7,9)),
          setequal(leaves(x), c(6,7,9)))

}
\author{
Magnus Dehli Vigeland
}
