% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedtools-package.R
\docType{package}
\name{pedtools-package}
\alias{pedtools}
\alias{pedtools-package}
\title{pedtools: Creating and Working with Pedigrees and Marker Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive collection of tools for creating, manipulating and visualising pedigrees and genetic marker data. Pedigrees can be read from text files or created on the fly with built-in functions. A range of utilities enable modifications like adding or removing individuals, breaking loops, and merging pedigrees. An online tool for creating pedigrees interactively, based on 'pedtools', is available at \url{https://magnusdv.shinyapps.io/quickped}. 'pedtools' is the hub of the 'ped suite', a collection of packages for pedigree analysis. A detailed presentation of the 'ped suite' is given in the book 'Pedigree Analysis in R' (Vigeland, 2021, ISBN:9780128244302).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/pedtools}
  \item \url{https://magnusdv.github.io/pedsuite/}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

}
\keyword{internal}
