% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedtools-package.R
\docType{package}
\name{pedtools}
\alias{pedtools}
\title{pedtools: Tools for working with pedigrees in R}
\description{
A comprehensive collection of tools for creating, manipulating and
visualising pedigrees and genetic marker data. Pedigrees can be read from
text files or created on the fly with built-in functions. A range of
utilities enable modifications like adding or removing individuals, breaking
loops, and merging pedigrees. Pedigree plots are produced by wrapping the
plotting functionality of the \code{kinship2} package. A Shiny app for creating
pedigrees, based on \code{pedtools}, is available at
\url{https://magnusdv.shinyapps.io/quickped}. \code{pedtools} is the hub of the \verb{ped suite}, a collection of packages for pedigree analysis. A detailed
presentation of the \verb{ped suite} is given in the book \href{https://www.elsevier.com/books/pedigree-analysis-in-r/vigeland/978-0-12-824430-2}{Pedigree Analysis in R}
(Vigeland, 2021, ISBN:9780128244302).
}
