% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPed.R
\name{readPed}
\alias{readPed}
\title{Read a pedigree from a set of files}
\usage{
readPed(prefix, what = NULL, verbose = TRUE)
}
\arguments{
\item{prefix}{a filepath without extension from which .ped, .map, .freq and
.dat files will be read.}

\item{what}{a subset of the vector c("ped", "map", "dat", "freq") indicating
which files to read. By default all files that are present will be read.}

\item{verbose}{wheteher to show descriptive messages while reading the
pedigree. Defaults to TRUE.}
}
\value{
a \code{ped} object
}
\description{
Assumptions:
\itemize{
\item All markers present in .dat, .freq and .map files are also present in the .ped file and in the same order.
\item The .dat file only contains a list of markers and is thus only used for validation.
}
}
\seealso{
\code{\link[=writePed]{writePed()}}
}
\author{
Elias Hernandis \href{mailto:eliashernandis@gmail.com}{eliashernandis@gmail.com}
}
