% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectMarkers.R
\name{selectMarkers}
\alias{selectMarkers}
\alias{getMarkers}
\alias{removeMarkers}
\alias{whichMarkers}
\title{Select or remove marker objects}
\usage{
selectMarkers(x, markeridx = NULL, markernames = NULL, chroms = NULL,
  fromPos = NULL, toPos = NULL)

getMarkers(x, markeridx = NULL, markernames = NULL, chroms = NULL,
  fromPos = NULL, toPos = NULL)

removeMarkers(x, markeridx = NULL, markernames = NULL, chroms = NULL,
  fromPos = NULL, toPos = NULL)

whichMarkers(x, markernames = NULL, chroms = NULL, fromPos = NULL,
  toPos = NULL)
}
\arguments{
\item{x}{A \code{ped} object}

\item{markeridx}{A numeric index vector specifying the markers to be
selected/removed.}

\item{markernames}{A character vector with marker names, or NULL}

\item{chroms}{An integer vector, or NULL}

\item{fromPos}{A single number or NULL}

\item{toPos}{A single number or NULL}
}
\value{
Each of \code{selectMarkers()} and \code{removeMarkers()} returns a \code{ped}
object, where the specified markers are attached/removed. The function
\code{getMarkers()} returns only the \code{markerList} object, while \code{whichMarkers()}
returns an integer vector with the indicies (position in \code{x$markerdata}) of
the specified markers. NULL arguments are skipped, so \code{whichMarkers(x)}
will return \code{seq_len(nMarkers(x))} (i.e. all markers).
}
\description{
Get the index (position in the markerdata list) of one or several markers
with specified properties.
}
\seealso{
\code{\link[=setMarkers]{setMarkers()}}
}
