% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedModify.R
\name{setFamid}
\alias{setFamid}
\title{Family ID}
\usage{
setFamid(x, famid)
}
\arguments{
\item{x}{a \code{ped} object}

\item{famid}{a character of length 1. If missing, an emtpy string is used.}
}
\description{
Set a family identifier of a \code{ped} object.
}
