% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_modify.R
\name{setLabels}
\alias{setLabels}
\title{Pedigree member labels}
\usage{
setLabels(x, labels)
}
\arguments{
\item{x}{a \code{ped} object}

\item{labels}{a character (or coercible to character) of length \code{x$NIND}}
}
\description{
Set labels of pedigree members
}
