% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMarkers.R
\name{setMarkers}
\alias{setMarkers}
\title{Attach marker objects to a pedigree}
\usage{
setMarkers(x, m, annotations = NULL)
}
\arguments{
\item{x}{A \code{ped} object}

\item{m}{Either a single \code{marker} object, a list of \code{marker} objects, or a data.frame or matrix.}

\item{annotations}{A list of marker annotations.}
}
\description{
Attach marker objects to a pedigree
}
