% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_utils.R
\name{subnucs}
\alias{subnucs}
\alias{print.nucleus}
\title{Nuclear subsets of a pedigree}
\usage{
subnucs(x)

\method{print}{nucleus}(x, ...)
}
\arguments{
\item{x}{A \code{ped} object.}

\item{...}{Not used.}
}
\value{
A list of \code{nucleus} objects. Each nucleus is a list with elements \code{father}, \code{mother} and \code{children}.
}
\description{
Nuclear subsets of a pedigree
}
\examples{
x = cousinsPed(0, child=TRUE)
subnucs(x)

}
