% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_transfer.R
\name{transferMarkers}
\alias{transferMarkers}
\title{Transfer marker data}
\usage{
transferMarkers(from, to)
}
\arguments{
\item{from}{a \code{ped} or \code{singleton} object, or a
list of such objects.}

\item{to}{a \code{ped} or \code{singleton} object, or a
list of such objects.}
}
\value{
A \code{ped} object (or a list of such) similar to \code{to}, but
where all individuals also present in \code{from} have marker genotypes
copied over.  Any previous marker data is erased.
}
\description{
Transfer marker data between pedigrees. Any markers attached to the
target are overwritten.
}
\examples{

x = nuclearPed(fa = "father", mo = "mother", children = "boy")
m = marker(x, father = 1:2, boy = 1:2)
x = setMarkers(x, m)

y = list(singleton("father"), nuclearPed(children = "boy"))
y = transferMarkers(x, y)
y
stopifnot(genotype(y[[1]], 1, "father") == 1:2, genotype(y[[2]], 1, "boy") == 1:2)

}
