% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped.R
\name{validate_ped}
\alias{validate_ped}
\title{Pedigree errors}
\usage{
validate_ped(x)
}
\arguments{
\item{x}{object of class \code{ped}.}
}
\value{
If no errors are detected, the function returns NULL invisibly.
Otherwise, messages describing the errors are printed to the screen and an
error is raised.
}
\description{
Validate the internal structure of a \code{ped} object.
}
