% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_utils.R
\name{is.marker}
\alias{is.marker}
\alias{is.markerList}
\title{Test if something is a marker}
\usage{
is.marker(x)

is.markerList(x)
}
\arguments{
\item{x}{Any object}
}
\value{
A logical
}
\description{
Functions for testing if something is a \code{marker} object, or a list of such objects.
}
