% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_attach.R
\name{marker_attach}
\alias{marker_attach}
\alias{setMarkers}
\alias{addMarkers}
\title{Attach markers to pedigrees}
\usage{
setMarkers(x, m = NULL, allele_matrix = NULL,
  locus_annotations = NULL, missing = 0, allele_sep = NULL)

addMarkers(x, m = NULL, allele_matrix = NULL,
  locus_annotations = NULL, missing = 0, allele_sep = NULL)
}
\arguments{
\item{x}{A \code{ped} object}

\item{m}{Either a single \code{marker} object or a list of \code{marker} objects}

\item{allele_matrix}{A matrix with \code{pedsize(x)} rows, containing the observed
alleles for one or several markers. The matrix must have either 1 or 2
columns per marker. If the former, then the argument \code{allele_sep} must be
non-NULL, and will be used to split all entries.}

\item{locus_annotations}{A list of lists, with annotations for each marker.
See \code{\link[=marker]{marker()}} for possible entries.}

\item{missing}{A single character (or coercible to one) indicating the symbol
for missing alleles.}

\item{allele_sep}{If this is a single character (instead of NULL), each entry
of \code{allele_matrix} is interpreted as a genotype, and will be split by
calling \code{str_split(..., split = allele_sep, fixed = T)}. For example, if
the entries are formatted as "A/B", put \code{allele_sep="/"}.}
}
\value{
A \code{ped} object.
}
\description{
In many applications it is useful to \emph{attach} markers to their associated
\code{ped} object. In particular for bigger projects with many markers, this makes
it easier to manipulate the dataset as a unit. The function \code{setMarkers()}
replaces all existing markers with the supplied ones, while \code{addMarkers()}
appends them to any existing ones.
}
\examples{
x = singleton(1)
m1 = marker(x, '1'=1:2)
m2 = marker(x, '1'='a')

x = setMarkers(x, m1)
x = addMarkers(x, m2)
x

# Reversing the order of the markers
x = setMarkers(x, list(m2, m1))
x

}
