% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_modify.R
\name{ped_modify}
\alias{ped_modify}
\alias{swapSex}
\alias{relabel}
\alias{setLabels}
\title{Pedigree modifications}
\usage{
swapSex(x, ids, verbose = TRUE)

relabel(x, new, old = labels(x))

setLabels(x, labels)
}
\arguments{
\item{x}{A \code{ped} object.}

\item{ids}{A character (or coercible to character) with ID labels of one or
more pedigree members.}

\item{verbose}{A logical: Verbose output or not.}

\item{new, old}{Character vectors (or coercible to character) of the same
length. ID labels in \code{old} are replaced by those in \code{new}.}

\item{labels}{A character (or coercible to character) of length \code{pedsize(x)}}
}
\value{
The modified \code{ped} object.
}
\description{
Functions for modifying various attributes of a 'ped' object.
}
\examples{

x = nuclearPed(1)

# To see the effect of each command below, use plot(x) in between.
x = swapSex(x, 3)
x = relabel(x, new="girl", old=3)

}
\seealso{
\code{\link[=ped]{ped()}}, \link{ped_add}
}
\author{
Magnus Dehli Vigeland
}
