% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_utils.R
\name{pedSize}
\alias{pedSize}
\title{Pedigree size}
\usage{
pedSize(x)
}
\arguments{
\item{x}{a \code{\link{ped}} object.}
}
\value{
A positive integer.
}
\description{
Return the number of pedigree members.
}
\examples{
x = nuclearPed(1)
stopifnot(pedSize(x)==3)

}
